changes step seq note duration to step / 2
this might not be important to anyone else, but this is useful to me
because I'm using gate cv to trigger drums, and adjacent steps were
keeping the gate open instead of retriggering the drum.  by setting the
note duration to step / 2 we ensure that the gate will always close
before the next step.  this probably breaks sample playback.  something
like this probably makes sense to move into configuration somewhere.